services/mcp-tasks/pyproject.toml
[project]
name = "mcp-tasks"
version = "0.2.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "celery[redis]>=5.4.0",
    "redis>=5.0.7",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "SQLAlchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "pymupdf>=1.24.2",
    "python-docx>=1.1.2"
]
